requirements.txt
Django==1.8.6
django-bootstrap3==6.2.2
django-extensions==1.5.9
django-redis==4.3.0
djangorestframework==3.3.1
gunicorn==19.3.0
Pillow==3.0.0
psycopg2==2.5.4
raven==5.1.0
redis==2.10.5
requests==2.8.1
six==1.10.0
sorl-thumbnail==12.3